Revert "add not found filter"
This reverts commit c03369ae1a5a2b5ec33a7a388e9ca6baa3118018.

+++ src/favorites/favorites.controller.ts
@@ -3,18 +3,26 @@ import {
   Delete,
   Get,
   HttpCode,
-  HttpStatus,
+  NotFoundException,
   Param,
   ParseUUIDPipe,
   Post,
-  UseFilters,
+  UnprocessableEntityException,
 } from '@nestjs/common';
-import { NotFoundExceptionFilter } from 'src/filters/entity-not-found-exception.filter';
+import { StatusCodes } from 'http-status-codes';
+import { AlbumService } from 'src/album/album.service';
+import { ArtistService } from 'src/artist/artist.service';
+import { TrackService } from 'src/track/track.service';
 import { FavoritesService } from './favorites.service';
 
 @Controller('favs')
 export class FavoritesController {
-  constructor(private readonly favoritesService: FavoritesService) {}
+  constructor(
+    private readonly favoritesService: FavoritesService,
+    private readonly trackService: TrackService,
+    private readonly albumService: AlbumService,
+    private readonly artistService: ArtistService,
+  ) {}
 
   @Get()
   findAll() {
@@ -22,41 +30,71 @@ export class FavoritesController {
   }
 
   @Post('track/:id')
-  @UseFilters(new NotFoundExceptionFilter(HttpStatus.UNPROCESSABLE_ENTITY))
   addTrackToFavorite(@Param('id', ParseUUIDPipe) id: string) {
+    const track = this.trackService.findOne(id);
+
+    if (!track) {
+      throw new UnprocessableEntityException('Track not found');
+    }
+
     this.favoritesService.addTrackToFavorite(id);
   }
 
   @Post('album/:id')
-  @UseFilters(new NotFoundExceptionFilter(HttpStatus.UNPROCESSABLE_ENTITY))
   addAlbumToFavorite(@Param('id', ParseUUIDPipe) id: string) {
+    const album = this.albumService.findOne(id);
+
+    if (!album) {
+      throw new UnprocessableEntityException('Album not found');
+    }
+
     this.favoritesService.addAlbumToFavorite(id);
   }
 
   @Post('artist/:id')
-  @UseFilters(new NotFoundExceptionFilter(HttpStatus.UNPROCESSABLE_ENTITY))
   addArtistToFavorite(@Param('id', ParseUUIDPipe) id: string) {
+    const artist = this.artistService.findOne(id);
+
+    if (!artist) {
+      throw new UnprocessableEntityException('Artist not found');
+    }
+
     this.favoritesService.addArtistToFavorite(id);
   }
 
   @Delete('track/:id')
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @UseFilters(new NotFoundExceptionFilter())
+  @HttpCode(StatusCodes.NO_CONTENT)
   removeTrackFromFavorite(@Param('id', ParseUUIDPipe) id: string) {
+    const track = this.trackService.findOne(id);
+
+    if (!track) {
+      throw new NotFoundException('Track not found');
+    }
+
     this.favoritesService.removeTrackFromFavorite(id);
   }
 
   @Delete('album/:id')
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @UseFilters(new NotFoundExceptionFilter())
+  @HttpCode(StatusCodes.NO_CONTENT)
   removeAlbumFromFavorite(@Param('id', ParseUUIDPipe) id: string) {
+    const album = this.albumService.findOne(id);
+
+    if (!album) {
+      throw new NotFoundException('Album not found');
+    }
+
     this.favoritesService.removeAlbumFromFavorite(id);
   }
 
   @Delete('artist/:id')
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @UseFilters(new NotFoundExceptionFilter())
+  @HttpCode(StatusCodes.NO_CONTENT)
   removeArtistFromFavorite(@Param('id', ParseUUIDPipe) id: string) {
+    const artist = this.artistService.findOne(id);
+
+    if (!artist) {
+      throw new NotFoundException('Artist not found');
+    }
+
     this.favoritesService.removeArtistFromFavorite(id);
   }
 }
